requirements.txt
tensorflow==2.16.1
streamlit
flask
firebase-admin
Pillow
pandas
requests
ipykernel
